Add files via upload
This file includes work done by me in week6, in lab 1, It includes some sort of coding, I learned it and practiced it using some examples in Lab 1, like creating some sounds of animals, or showing a starting engine of a car, and its battery level, etc. I tried doing some changes to it, such as changing car brand and battery percentage.